Bug fixes.
--HG--
branch : dev-css

+++ src/eepp/ui/uilinearlayout.cpp
@@ -97,7 +97,7 @@ void UILinearLayout::packVertical() {
 				{
 					int w = mDpSize.getWidth() - widget->getLayoutMargin().Left - widget->getLayoutMargin().Right - mPadding.Left - mPadding.Right;
 
-					if ( widget->getSize().getWidth() != w )
+					if ( (int)widget->getSize().getWidth() != w )
 						widget->setSize( w, widget->getSize().getHeight() );
 
 					break;
@@ -166,14 +166,14 @@ void UILinearLayout::packVertical() {
 	if ( getLayoutHeightRules() == WRAP_CONTENT ) {
 		curY += mPadding.Bottom;
 
-		if ( curY != mDpSize.getHeight() ) {
+		if ( curY != (int)mDpSize.getHeight() ) {
 			setInternalHeight( curY );
 			notifyLayoutAttrChangeParent();
 		}
 	} else if ( getLayoutHeightRules() == MATCH_PARENT ) {
 		int h = getParent()->getSize().getHeight() - mLayoutMargin.Top - mLayoutMargin.Bottom;
 
-		if ( h != mDpSize.getHeight() )
+		if ( h != (int)mDpSize.getHeight() )
 			setInternalHeight( h );
 	}
 
@@ -284,14 +284,14 @@ void UILinearLayout::packHorizontal() {
 	if ( getLayoutWidthRules() == WRAP_CONTENT ) {
 		curX += mPadding.Right;
 
-		if ( curX != mDpSize.getWidth() ) {
+		if ( curX != (int)mDpSize.getWidth() ) {
 			setInternalWidth( curX );
 			notifyLayoutAttrChangeParent();
 		}
 	} else if ( getLayoutWidthRules() == MATCH_PARENT ) {
 		int w = getParent()->getSize().getWidth() - mLayoutMargin.Left - mLayoutMargin.Right;
 
-		if ( w != mDpSize.getWidth() )
+		if ( w != (int)mDpSize.getWidth() )
 			setInternalWidth( w );
 	}
